updating with transition timing attempt
Do not merge this to main. this transition timing does not work because this is a data-driven paint property. If we want to implement this as a feature we're going to need to get a hard coded layer attribute an change it in a way that makes the layer invisible, switch the styling and then make it visible. Deprioritizing for now, but keeping branch open if others have better ideas.